Create Sep31DepositInfoGeneratorCircle class in platform.
